chore: move code to be re-used to `support.rs`
